Moved permission checking to host activity
